Use global-phase ranking and remove java customizations
