+++ transformers/application/search/query-profiles/default.xml
@@ -1,0 +1,5 @@
+<query-profile id="default">
+    <field name="maxHits">100</field>
+    <field name="maxOffset">1000</field>
+    <field name="timeout">5s</field>
+</query-profile>
