+++ transformers/application/services.xml
@@ -1,0 +1,44 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!-- Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
+<services version="1.0" xmlns:deploy="vespa" xmlns:preprocess="properties">
+
+  <container id="default" version="1.0">
+    <document-api/>
+
+    <component id="tokenizer"
+               class="com.yahoo.language.wordpiece.WordPieceEmbedder" bundle="linguistics-components">
+      <config name="language.wordpiece.word-piece">
+        <model>
+          <item>
+            <language>unknown</language>
+            <path>files/bert-base-uncased-vocab.txt</path>
+          </item>
+        </model>
+      </config>
+    </component>
+    <search/>
+  </container>
+
+
+  <content id='msmarco' version='1.0'>
+    <redundancy>2</redundancy>
+    <documents>
+      <document type='msmarco' mode="index"/>
+    </documents>
+    <nodes>
+      <node hostalias="node1" distribution-key="0" />
+    </nodes>
+    <engine>
+      <proton>
+        <tuning>
+          <searchnode>
+            <requestthreads>
+              <persearch>1</persearch>
+            </requestthreads>
+          </searchnode>
+        </tuning>
+      </proton>
+    </engine>
+  </content>
+
+</services>
