--- transformers/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-
-<!-- Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0                              http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>ai.vespa.examples</groupId>
-  <artifactId>transformers</artifactId>
-  <packaging>container-plugin</packaging>
-  <version>1.0.0</version>
-
-  <parent>
-    <groupId>com.yahoo.vespa</groupId>
-    <artifactId>cloud-tenant-base</artifactId>
-    <version>[8,9)</version>  <!-- Use the latest Vespa release on each build -->
-    <relativePath/>
-  </parent>
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <test.hide>true</test.hide>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>com.yahoo.vespa</groupId>
-      <artifactId>linguistics-components</artifactId>
-      <scope>provided</scope>
-      <version>[8,9)</version>
-    </dependency>
-  </dependencies>
-
-</project>
