--- transformers/src/main/application/search/query-profiles/default.xml
@@ -1,7 +1,0 @@
-<query-profile id="default">
-    <field name="maxHits">100</field>
-    <field name="maxOffset">1000</field>
-    <field name="timeout">60s</field>
-    <field name="searchChain">transformer</field>
-    <field name="ranking.queryCache">true</field>
-</query-profile>
