--- transformers/src/main/application/services.xml
@@ -1,58 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!-- Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
-<services version="1.0" xmlns:deploy="vespa" xmlns:preprocess="properties">
-
-  <container id="default" version="1.0">
-    <document-api/>
-
-    <component id="com.yahoo.language.wordpiece.WordPieceEmbedder" bundle="linguistics-components">
-      <config name="language.wordpiece.word-piece">
-        <model>
-          <item>
-            <language>unknown</language>
-            <path>files/bert-base-uncased-vocab.txt</path>
-          </item>
-        </model>
-      </config>
-    </component>
-
-    <document-processing>
-      <chain id="transformer-processing" inherits="indexing">
-        <documentprocessor id="ai.vespa.examples.processor.TransformerProcessor" bundle="transformers"/>
-      </chain>
-    </document-processing>
-
-    <search>
-      <chain id="transformer" inherits="vespa">
-        <searcher id="ai.vespa.examples.searcher.TransformerSearcher" bundle="transformers"/>
-      </chain>
-    </search>
-  </container>
-
-
-  <content id='msmarco' version='1.0'>
-    <search>
-      <visibility-delay>1</visibility-delay>
-    </search>
-    <redundancy>2</redundancy>
-    <documents>
-      <document type='msmarco' mode="index"/>
-      <document-processing cluster="default" chain="transformer-processing" />
-    </documents>
-    <nodes>
-      <node hostalias="node1" distribution-key="0" />
-    </nodes>
-    <engine>
-      <proton>
-        <tuning>
-          <searchnode>
-            <requestthreads>
-              <persearch>10</persearch>
-            </requestthreads>
-          </searchnode>
-        </tuning>
-      </proton>
-    </engine>
-  </content>
-
-</services>
