--- transformers/src/main/resources/configdefinitions/vocabulary.def
@@ -1,5 +1,0 @@
-# Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-
-package=ai.vespa.tokenizer
-
-vocabulary path
