--- transformers/src/test/application/tests/sample-document.json
@@ -1,9 +1,0 @@
-{
-  "put": "id:msmarco:msmarco::D2779191",
-  "fields": {
-    "id": "D2779191",
-    "title": "Is Long Term Care Insurance Tax Deductible  2016 Tax Year Updated",
-    "url": "https://www.ltctree.com/long-term-care-insurance-tax-deductible/",
-    "body": "Last Updated March 15  2017Is Long Term Care Insurance Tax Deductible  2016 Tax Year Updated Long Term Care Insurance policies sold are either IRS tax qualified or non tax qualified however  most policies sold today are tax qualified"
-  }
-}
